+++ CHANGELOG.md
@@ -12,3 +12,4 @@ Changelog
 
 ## Fix
 - [PADA-216](https://givisiez.atlassian.net/browse/PADA-216): Fix conflict and import errors
+- [PADA-209](https://givisiez.atlassian.net/browse/PADA-216): Modified MyVaccines screen, components and style